Interaction, movement and engine rewrite (#320)
* Move OverlapTest

* Add more missing packages

* Move koin extensions

* Fix experience test

* Fix up dialogue tests

* Add Movement mode unit tests

* Fix ChunkBatches unit tests

* Add reified npc statement expressions

* Add reified player statement expressions

* Convert expressions in tests

* Convert one use timers to queue's

* Convert World timers

* Refactor Timers

* Add string names to all timers

* Remove unused visible entity methods

* Add names

* Rename Time to Ticks

* Add events to timers

* Stop returning Timer's

* Add timer tick event

* Renaming timers

* Convert some content to timers

* Add Clocks

* Convert timers to count down and be Event based

* Convert antifire and overloads

* Add cancellable timer starting

* Convert poison to timer

* Convert god bows

* Convert ava's

* Convert colour overlays

* Add custom tick for World tasks

* Add diseases

* Convert game tick to int

* Ticks

* Convert some clocks

* Add transmit flag to variables

* Add delegate to clocks

* Add custom variables for all existing clocks; freeze, transform, prayers, combat

* Convert and remove effects

* Remove longs from toTicks

* Convert some persistent clocks to variables

* Store bitwise variables as Set's

* Convert skill stat's to snake case

* Convert prayers to snake case

* Fix world timers

* Fix level up dialogue

* Fix action queue not being removed

* Add string names to all queues for debuggability

* Extract VariableBits from Variables

* Rename add to set

* Simplify and refactor variables

* Split PlayerVariables from Variables

* Make variable clocks more general

* Replace Clocks with Variables extensions

* Convert Values

* Delete clock delegates

* Remove Values

* Refactoring variable functions

* Split out variable clock functions

* Refactor praying

* Convert in wilderness to vars

* Commands

* Return result for bitwise changes

* Convert music tracks to use string id's

* Prayers

* Handle custom variable's better

* Load different variable types from individual files

* Don't save character creation containers

* Remove custom variable formats

* Tidy up variable types

* Save players on main thread to avoid VariableData.persist threading issues

* Experimental interaction state

* Change how interaction works

* Add skill delays

* Tidy up player login and logout methods

* Revert interact change

* Clear watch on clear interact

* Tweaks to firemaking, still not quite correct; no bouncing off, turning isn't correct, spam lighting different

* Stop bone burying cancelling movement

* Slightly better firemaking movement and turning

* Add one item cooking

* Add instant initial cooking delay

* Fix item dropping interrupting fishing

* Fix fishing animation being cleared

* Fix skill creation making 1 dialogue

* Close dialogue on interaction

* Fix duplicate chunk updates being hashed together

* Fix 1 and 2 tick cooking

* Increase Delta 8k x/y limit to 32k

* Small update task refactor

* Fix firemaking interaction

* Add firemaking bouncing off objects

* Remove overly aggressive interaction clearing

* Close interfaces when appropriate

* Fix dressing room

* Dropping an item doesn't stop movement

* Clear interfaces in the correct places

* Clear weak queue on interface close attempts

* Remove StopInteraction.kt

* Tweaks

* Don't close bank when item switching

* Turn to face player when interacting

* Refactor Interact.kt

* Fix viewport not loading race condition

* Whoops

* Fix trade and assist request messages

* Add interact tests

* Fix interaction and setting approach range for bankers

* Add approach or operate interactions to all npc options

* Improve item pickup delays

* Fix facing entities

* Add operation to all items on objects

* Add approach distances to spell on others

* Close interfaces on tele

* Fix object interaction bitmask

* Add stair interaction delay

* Fix teleporting delays

* Fix resting

* Remove duplicate exception handlers

* Properly deallocate cache, and general memory reduction ~82MB

* Fix removing transformation command

* Fix block animations

* Extract target strategy target character

* Fix transform

* Add basic combat mode

* Fix npc death

* Fix combat movement

* Fix spell auto casting

* Fix npc hitpoints

* Fix giant rat npc stats

* Fix npc death

* Fix npc health bar

* Add npc respawn delays

* Fix npc respawn delay

* Only make npcs invisible when dead

* Fix combat interact range, standing under target and long range weapons

* Remove print

* Fix interacting not clearing dialogues after finished

* Fix floor item stack combining

* Fix engine unit tests

* Tweak extractor for speed

* Only non-dialogue interfaces remove weak queues

* Clear weak queues on switching and dropping items

* Convert item on item to use queue's instead of pause()

* Fix level calculation to handle decimals

* Fix archery target

* Split dialogue and regular suspensions, so they don't interfere with each other

* Fix range tutor dialogue

* Fix walking not clearing dialogues

* Fix combat dummies

* Fix timer clearing on death

* Stop combat on inventory actions

* Fix container sizes

* Fix lumbridge fishing shop samples

* Fix shop stock messages

* Fix item info by always sending clearing player variables

* Fix object interaction direction

* Fix door interaction

* Clear weak queues on dialogue clearing

* Fix dijkstra unit tests

* Remove VariableData

* Update PathFinder version to fix walkable out of bounds

* Correctly allocate collision map chunks

* Fix collision clearing

* Remove old comments

* Add better area collision checking for tele tabs

* Tidy up account loader

* Clear xp counter on reset command

* Fix talking with Thessalia

* Add movement delay to emotes

* Don't clear suspension on walk to prevent tele queue looping

* Move steps out of Movement

* Change Tollgate message

* Add better temporary walking, auto move when attempt to walk into middle of BorderGuards

* Fix al-kharid tollgate

* Remove unused forced movement

* Tollgate stuff

* Prevent combat hitting after death

* Fix combat movement

* Variable debug command

* Fix single cast spells

* Fix player variable lookup

* Fix interaction movement

* Add npc wander radius

* Fix mining loop

* Fix player x-log

* Add giant rat animations

* Update rsmod pf library version

* Fix re-adding player bot

* Fix bot movement

* Fix unit test compiling

* Add missing expression

* Clear suspensions on completion

* Fix following & watching

* Fix bone burying delay

* Tweaks; fire making, picking up item logs

* Fix broken floor item state

* Fix mining & display mode changing

* Remove double player context from await interfaces

* Clear animation on mining ore success

* Cancel npc suspensions on change too

* Fix combat movement

* Client tweak

* Fix movement calculation

* Clear steps on death

* Add stepping out from under target and fix calc again

* Add npc death delay

* Fix integration tests

+++ game/src/main/kotlin/world/gregs/voidps/bot/navigation/Navigation.kt
@@ -3,14 +3,19 @@ package world.gregs.voidps.bot.navigation
 import kotlinx.coroutines.CancellableContinuation
 import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.coroutines.withTimeoutOrNull
-import world.gregs.voidps.engine.entity.*
-import world.gregs.voidps.engine.entity.character.player.Bot
+import world.gregs.voidps.bot.Bot
+import world.gregs.voidps.engine.client.variable.contains
+import world.gregs.voidps.engine.client.variable.remove
+import world.gregs.voidps.engine.entity.Entity
+import world.gregs.voidps.engine.client.variable.get
+import world.gregs.voidps.engine.client.variable.getOrPut
+import world.gregs.voidps.engine.client.variable.set
 import world.gregs.voidps.engine.event.Event
-import world.gregs.voidps.engine.utility.TICKS
+import world.gregs.voidps.engine.timer.TICKS
 import kotlin.coroutines.resume
 import kotlin.reflect.KClass
 
-suspend fun Bot.await(type: Any, timeout: Long = -1) {
+suspend fun Bot.await(type: Any, timeout: Int = -1) {
     if (timeout > 0) {
         withTimeoutOrNull(TICKS.toMillis(timeout)) {
             suspendCancellableCoroutine<Unit> { cont ->
